requirements.txt
streamlit
boto3
python-pptx
requests
pandas